fixing routes

+++ routes/api.php
@@ -44,10 +44,7 @@ Route::middleware(JwtMiddleware::class)->group(function () {
     Route::get('notifications/tokens', [NotificationController::class, 'getUserTokens']);
     Route::post('/notifications/token/unregister', [NotificationController::class, 'unregisterToken']);
 
-
-    Route::middleware('is_admin')->prefix('admin')->group(function () {
-        Route::get('users/roles', [AuthController::class, 'getUserRoles']);
-        Route::put('users/{id}/role', [AuthController::class, 'updateUserRole']);
-        Route::delete('users/{id}', [AuthController::class, 'deleteUser']);
-    });
+    Route::get('admin/users', [AuthController::class, 'getUserRoles']);
+    Route::put('admin/users/{id}/role', [AuthController::class, 'updateUserRole']);
+    Route::delete('admin/users/{id}', [AuthController::class, 'deleteUser']);
 });
